requirements.txt
fastapi
uvicorn
tensorflow
opencv-python
numpy
pandas
pillow
scikit-image
pandas
